[layouts] move 'Development Artifacts' layouts into dev-layouts.
