+++ lib/js/components/dev-layouts/videoproof-array.mjs
@@ -48,7 +48,7 @@ import {
 import {
     initAnimanion
   , UITimeControl
-} from './stage-and-actors.mjs';
+} from '../layouts/stage-and-actors.mjs';
 
 
 const VideoproofArrayActorModel = activatableActorTypes.get('VideoproofArrayActorModel').get('typeClass').value;
