+++ lib/js/components/layouts/stage-and-actors.mjs
@@ -63,7 +63,7 @@ import {
 
 import {
     createLabelForKeyMoment
-} from './example-key-moments.mjs';
+} from '../dev-layouts/example-key-moments.mjs';
 
 import {
     UIManualAxesLocations
