+++ lib/js/components/main-ui.mjs
@@ -29,11 +29,11 @@ import {
   , collapsibleMixin
 } from './generic.mjs';
 
-import * as ExampleLayout from './layouts/example.mjs';
-import * as ExampleKeyMomentsLayout from './layouts/example-key-moments.mjs';
-import * as PentrekLayout from './layouts/pentrek.mjs';
+import * as ExampleLayout from './dev-layouts/example.mjs';
+import * as ExampleKeyMomentsLayout from './dev-layouts/example-key-moments.mjs';
+import * as PentrekLayout from './dev-layouts/pentrek.mjs';
 import * as StageAndActorsLayout from './layouts/stage-and-actors.mjs';
-import * as VideoproofArrayLayout from './layouts/videoproof-array.mjs';
+import * as VideoproofArrayLayout from './dev-layouts/videoproof-array.mjs';
 import * as VideoproofArrayV2Layout from './layouts/videoproof-array-v2.mjs';
 import * as TypeSpecRamp from './layouts/type-spec-ramp.mjs';
 import * as TypeToolsGrid from './layouts/type-tools-grid.mjs';
